requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
idna==3.3
mysqlclient==2.1.0
Pillow==8.4.0
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
